email_verifier_fastapi/requirements.txt
fastapi
uvicorn
dnspython
smtplib
cachetools
uvicorn
cachetools
requests
